requirements.txt
# Procesamiento de Datos [cite: 11, 12]
pandas>=2.0.0
numpy>=1.24.0
biopython>=1.81

# Inteligencia Artificial y Machine Learning [cite: 13]
scikit-learn>=1.3.0
tensorflow>=2.13.0
# pytorch lo dejamos opcional por ahora para no hacerlo muy pesado

# Simulación Científica [cite: 14]
scipy>=1.11.0
matplotlib>=3.7.0

# Visualización Web / Dashboard [cite: 16]
dash>=2.11.0
plotly>=5.15.0

# Utilidades
requests>=2.31.0